ref: Move path parsing code
